fix mistakes and rename module
