+++ src/main/java/me/moomoo/anarchyexploitfixes/AnarchyExploitFixes.java
@@ -4,7 +4,7 @@ import me.moomoo.anarchyexploitfixes.commands.AnarchyExploitFixesCommand;
 import me.moomoo.anarchyexploitfixes.config.Config;
 import me.moomoo.anarchyexploitfixes.config.LanguageCache;
 import me.moomoo.anarchyexploitfixes.modules.AnarchyExploitFixesModule;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.ProtocolLibModule;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -165,7 +165,7 @@ public class AnarchyExploitFixes extends JavaPlugin {
     private void reloadConfiguration() {
         config = new Config();
         AnarchyExploitFixesModule.reloadModules();
-        AEFProtocolLibModule.reloadModules();
+        ProtocolLibModule.reloadModules();
         config.saveConfig();
     }
