--- src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/AEFProtocolLibModule.java
@@ -1,43 +1,0 @@
-package me.moomoo.anarchyexploitfixes.modules.protocollib;
-
-import com.comphenix.protocol.ProtocolLibrary;
-import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.boatfly.AntiBoatFlyModule;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.craftrecipe.CraftRecipeModule;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.nocom.NoComModule;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.packetfly.AntiPacketFlyModule;
-
-import java.util.HashSet;
-
-public interface AEFProtocolLibModule {
-
-    String name();
-    String category();
-    void enable();
-    boolean shouldEnable();
-
-    HashSet<AEFProtocolLibModule> modules = new HashSet<>();
-
-    static void reloadModules() {
-        modules.clear();
-        if (AnarchyExploitFixes.isProtocolLibInstalled()) {
-            AnarchyExploitFixes plugin = AnarchyExploitFixes.getInstance();
-            ProtocolLibrary.getProtocolManager().removePacketListeners(plugin);
-
-            modules.add(new NoComModule());
-            modules.add(new AntiBoatFlyModule());
-            modules.add(new AntiPacketFlyModule());
-            modules.add(new CraftRecipeModule());
-
-            for (AEFProtocolLibModule module : modules) {
-                if (module.shouldEnable()) module.enable();
-                if (module.name() != null) plugin.enabledModules.put("<" + module.category() + "> " + module.name(), module.shouldEnable());
-            }
-        } else {
-            if (!AnarchyExploitFixes.getConfiguration().protocolLib_IsDisabled) {
-                AnarchyExploitFixes.getLog().severe("Could not find ProtocolLib. Many gamebreaking exploits cannot be patched without it.");
-                AnarchyExploitFixes.getLog().severe("Download at: https://www.spigotmc.org/resources/protocollib.1997/");
-            }
-        }
-    }
-}
