+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/ProtocolLibModule.java
@@ -1,0 +1,44 @@
+package me.moomoo.anarchyexploitfixes.modules.protocollib;
+
+import com.comphenix.protocol.ProtocolLibrary;
+import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.boatfly.AntiBoatFlyModule;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.craftrecipe.CraftRecipeModule;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.nocom.NoComModule;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.packetfly.AntiPacketFlyModule;
+
+import java.util.HashSet;
+
+public interface ProtocolLibModule {
+
+    String name();
+    String category();
+    void enable();
+    boolean shouldEnable();
+
+    HashSet<ProtocolLibModule> modules = new HashSet<>();
+
+    static void reloadModules() {
+        modules.clear();
+        AnarchyExploitFixes plugin = AnarchyExploitFixes.getInstance();
+        ProtocolLibrary.getProtocolManager().removePacketListeners(plugin);
+
+        modules.add(new NoComModule());
+        modules.add(new AntiBoatFlyModule());
+        modules.add(new AntiPacketFlyModule());
+        modules.add(new CraftRecipeModule());
+
+        for (ProtocolLibModule module : modules) {
+            if (module.shouldEnable()) module.enable();
+            plugin.enabledModules.put("<" + module.category() + "> " + module.name(), module.shouldEnable());
+        }
+
+        if (
+                !AnarchyExploitFixes.isProtocolLibInstalled()
+                && !AnarchyExploitFixes.getConfiguration().protocolLib_IsDisabled
+        ) {
+            AnarchyExploitFixes.getLog().severe("Could not find ProtocolLib. Many gamebreaking exploits cannot be patched without it.");
+            AnarchyExploitFixes.getLog().severe("Download at: https://www.spigotmc.org/resources/protocollib.1997/");
+        }
+    }
+}
