+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/boatfly/AntiBoatFlyModule.java
@@ -2,16 +2,16 @@ package me.moomoo.anarchyexploitfixes.modules.protocollib.boatfly;
 
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
 import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.ProtocolLibModule;
 import me.moomoo.anarchyexploitfixes.utils.LogUtils;
 
 import java.util.logging.Level;
 
-public class AntiBoatFlyModule implements AEFProtocolLibModule {
+public class AntiBoatFlyModule implements ProtocolLibModule {
 
     private final Config config;
     private final int maxEntityPacketsPer10s;
-    private boolean logIsEnabled, shouldKickPlayer;
+    private final boolean logIsEnabled, shouldKickPlayer;
 
     public AntiBoatFlyModule() {
         this.config = AnarchyExploitFixes.getConfiguration();
@@ -43,8 +43,9 @@ public class AntiBoatFlyModule implements AEFProtocolLibModule {
                 LogUtils.moduleLog(Level.WARNING, name(), "Not patching exploit because you disabled ProtocolLib in config!");
                 return false;
             }
-            if (AnarchyExploitFixes.isProtocolLibInstalled()) {
+            if (!AnarchyExploitFixes.isProtocolLibInstalled()) {
                 LogUtils.moduleLog(Level.SEVERE, name(), "Unable to patch exploit because ProtocolLib is not installed!");
+                return false;
             }
             return true;
         }
