+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/boatfly/BoatFlyPatch.java
@@ -6,8 +6,6 @@ import com.comphenix.protocol.events.ListenerPriority;
 import com.comphenix.protocol.events.PacketAdapter;
 import com.comphenix.protocol.events.PacketEvent;
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
 import me.moomoo.anarchyexploitfixes.utils.LogUtils;
 import org.bukkit.entity.Boat;
 import org.bukkit.entity.Entity;
