+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/craftrecipe/CraftRecipeModule.java
@@ -2,12 +2,12 @@ package me.moomoo.anarchyexploitfixes.modules.protocollib.craftrecipe;
 
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
 import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.ProtocolLibModule;
 import me.moomoo.anarchyexploitfixes.utils.LogUtils;
 
 import java.util.logging.Level;
 
-public class CraftRecipeModule implements AEFProtocolLibModule {
+public class CraftRecipeModule implements ProtocolLibModule {
 
     private final Config config;
     private final long craftingRecipeDelayInMillis;
@@ -40,8 +40,9 @@ public class CraftRecipeModule implements AEFProtocolLibModule {
                 LogUtils.moduleLog(Level.WARNING, name(), "Not patching exploit because you disabled ProtocolLib in config!");
                 return false;
             }
-            if (AnarchyExploitFixes.isProtocolLibInstalled()) {
+            if (!AnarchyExploitFixes.isProtocolLibInstalled()) {
                 LogUtils.moduleLog(Level.SEVERE, name(), "Unable to patch exploit because ProtocolLib is not installed!");
+                return false;
             }
             return true;
         }
