+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/nocom/NoComBlockDigPacket.java
@@ -9,12 +9,7 @@ import com.comphenix.protocol.events.PacketEvent;
 import com.comphenix.protocol.wrappers.BlockPosition;
 import com.comphenix.protocol.wrappers.EnumWrappers;
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
-import me.moomoo.anarchyexploitfixes.utils.LogUtils;
 import org.bukkit.entity.Player;
-
-import java.util.logging.Level;
 
 public class NoComBlockDigPacket extends PacketAdapter {
