+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/nocom/NoComModule.java
@@ -2,12 +2,12 @@ package me.moomoo.anarchyexploitfixes.modules.protocollib.nocom;
 
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
 import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.ProtocolLibModule;
 import me.moomoo.anarchyexploitfixes.utils.LogUtils;
 
 import java.util.logging.Level;
 
-public class NoComModule implements AEFProtocolLibModule {
+public class NoComModule implements ProtocolLibModule {
 
     private final Config config;
     private final int noComMaxDistance;
@@ -41,8 +41,9 @@ public class NoComModule implements AEFProtocolLibModule {
                 LogUtils.moduleLog(Level.WARNING, name(), "Not patching exploit because you disabled ProtocolLib in config!");
                 return false;
             }
-            if (AnarchyExploitFixes.isProtocolLibInstalled()) {
+            if (!AnarchyExploitFixes.isProtocolLibInstalled()) {
                 LogUtils.moduleLog(Level.SEVERE, name(), "Unable to patch exploit because ProtocolLib is not installed!");
+                return false;
             }
             int mcVersion = AnarchyExploitFixes.getMCVersion();
             if (mcVersion == 17) {
