+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/nocom/NoComUseItemPacket.java
@@ -7,7 +7,6 @@ import com.comphenix.protocol.events.PacketAdapter;
 import com.comphenix.protocol.events.PacketEvent;
 import com.comphenix.protocol.wrappers.BlockPosition;
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
 import org.bukkit.entity.Player;
 
 public class NoComUseItemPacket extends PacketAdapter {
