+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/packetfly/AntiPacketFlyModule.java
@@ -2,12 +2,12 @@ package me.moomoo.anarchyexploitfixes.modules.protocollib.packetfly;
 
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
 import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
+import me.moomoo.anarchyexploitfixes.modules.protocollib.ProtocolLibModule;
 import me.moomoo.anarchyexploitfixes.utils.LogUtils;
 
 import java.util.logging.Level;
 
-public class AntiPacketFlyModule implements AEFProtocolLibModule {
+public class AntiPacketFlyModule implements ProtocolLibModule {
 
     private final Config config;
     private final boolean logIsEnabled, shouldKickPlayer;
@@ -43,8 +43,9 @@ public class AntiPacketFlyModule implements AEFProtocolLibModule {
                 LogUtils.moduleLog(Level.WARNING, name(), "Not patching exploit because you disabled ProtocolLib in config!");
                 return false;
             }
-            if (AnarchyExploitFixes.isProtocolLibInstalled()) {
+            if (!AnarchyExploitFixes.isProtocolLibInstalled()) {
                 LogUtils.moduleLog(Level.SEVERE, name(), "Unable to patch exploit because ProtocolLib is not installed!");
+                return false;
             }
             return true;
         }
