+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/protocollib/packetfly/PacketFlyPatch.java
@@ -7,8 +7,6 @@ import com.comphenix.protocol.events.PacketAdapter;
 import com.comphenix.protocol.events.PacketEvent;
 import com.cryptomorin.xseries.XMaterial;
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.modules.protocollib.AEFProtocolLibModule;
 import me.moomoo.anarchyexploitfixes.utils.LogUtils;
 import org.bukkit.Material;
 import org.bukkit.block.BlockFace;
